EngineLayer 프로젝트 디버그 모드 경고 레벨 4로 설정. 경고 나오지 않도록 #pragma warning 사용

+++ EngineLayer/SFPacketIOBuffer.cpp
@@ -44,7 +44,7 @@ bool SFPacketIOBuffer::GetPacket(SFPacketHeader& header, char* pBuffer, unsigned
 
 	int getDataSize = GetData((char*)pBuffer, dataSize);
 
-	if (dataSize != getDataSize)
+	if (dataSize != (unsigned int)getDataSize)
 	{
 		SetHead(oldHead);
 		errorCode = PACKETIO_ERROR_DATA;
